Extract (without reading .exe files) the icon stored on Windows Registry, for Local PC apps
