+++ src/UniGetUI.PackageEngine.PackageLoader/DiscoverablePackagesLoader.cs
@@ -1,4 +1,3 @@
-using UniGetUI.Core.Logging;
 using UniGetUI.Core.Tools;
 using UniGetUI.Interface.Enums;
 using UniGetUI.PackageEngine.Interfaces;
